+++ attributes/packages.rb
@@ -9,3 +9,7 @@ default['nodejs']['packages'] = value_for_platform_family(
   'freebsd' => %w(node npm),
   'default' => ['nodejs']
 )
+
+# Add options and actions by package name
+default['nodejs']['package_action'] = { default: :install }
+default['nodejs']['package_options'] = { default: '' }