+++ test/integration/default/default_spec.rb
@@ -1,3 +1,4 @@
+
 control 'commands should exist' do
   describe command('node -v') do
     its('exit_status') { should eq 0 }
@@ -7,19 +8,12 @@ control 'commands should exist' do
     its('exit_status') { should eq 0 }
   end
 
-  describe file '/home/random/staging' do
-    it { should exist }
-  end
-
   describe command('npx -v') do
     its('exit_status') { should eq 0 }
   end
 
-  describe file('/usr/local/bin/npm') do
-    it { should exist }
-  end
-
-  describe file('/usr/local/bin/npx') do
-    it { should exist }
+  describe command('npm list -json -global') do
+    its('stdout') { should_not match(/"express"/) }
+    its('stdout') { should match(/"socket.io"/) }
   end
 end