+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Add rspec tests for the library methods
 - Update the url_invalid? method to return false if it detects and invalid uri
 - Add the auto_update option to the npm_package resource. Allows turning off auto_update of npm packages.
+- Allow actions and options for OS package installation to be specified as attributes
+- Add the live_stream parameter to the npm_package execution to get better installation diagnostics
+- Add the auto_update option to the npm_package resource. Allows turning off auto_update of npm packages.
+- Update testing
 
 ## 7.2.0 (2020-10-07)